separate blobs plugin into the hard parts (api) from the soft parts (replication strategy). The hard parts are easy to test. The soft parts are the replication strategy, which, in a peer to peer system necessarily must handle failures and missing peers and works eventually... which makes it a bit messy to test, unfortunately.
